NODE-323: Moved gossiping under comm, leaving deployment, diagnostics and admin under node/api.
