+++ protobuf/io/casperlabs/comm/discovery/kademlia.proto
@@ -1,14 +1,7 @@
 syntax = "proto3";
 package io.casperlabs.comm.discovery;
 
-import "scalapb/scalapb.proto";
 import "io/casperlabs/comm/discovery/node.proto";
-
-option (scalapb.options) = {
-  package_name: "io.casperlabs.comm.discovery"
-  flat_package: true
-};
-
 
 message Ping {
   Node   sender         = 1;
