+++ protobuf/io/casperlabs/comm/discovery/node.proto
@@ -2,13 +2,6 @@ syntax = "proto3";
 
 package io.casperlabs.comm.discovery;
 
-import "scalapb/scalapb.proto";
-
-option (scalapb.options) = {
-  package_name: "io.casperlabs.comm.discovery"
-  flat_package: true
-};
-
 message Node {
     bytes id = 1;
     string host = 2;
