+++ protobuf/io/casperlabs/comm/protocol/routing/routing.proto
@@ -2,7 +2,6 @@ syntax = "proto3";
 package io.casperlabs.comm.protocol.routing;
 
 import "scalapb/scalapb.proto";
-import "io/casperlabs/casper/protocol/CasperMessage.proto";
 
 option (scalapb.options) = {
   package_name: "io.casperlabs.comm.protocol.routing"
