+++ protobuf/io/casperlabs/node/api/admin.proto
@@ -1,0 +1,12 @@
+syntax = "proto3";
+
+package io.casperlabs.node.api.admin;
+
+service AdminService {
+    // While we have the manual proposal option.
+    rpc Propose(ProposeRequest) returns (ProposeResponse) {}
+}
+
+message ProposeRequest {}
+
+message ProposeResponse {}
