+++ protobuf/io/casperlabs/node/api/diagnostics.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package io.casperlabs.node.model;
+package io.casperlabs.node.api.diagnostics;
 
 import "google/protobuf/empty.proto";
 import "google/protobuf/wrappers.proto";
